+++ assets/javascripts/kitten/components/layout/hero-layout/stories.js
@@ -1,0 +1,427 @@
+import React from 'react'
+import { HeroLayout } from './index'
+import { DocsPage } from 'storybook/docs-page'
+import styled, { createGlobalStyle } from 'styled-components'
+import {
+  Text,
+  HeaderNav,
+  KissKissBankBankLogo,
+  KissKissLoadingAnimation,
+  HEADER_HEIGHT,
+  pxToRem,
+  FlexWrapper,
+  Button,
+  Title,
+  mq,
+  TYPOGRAPHY,
+  Grid,
+  GridCol,
+  HeartIconNext,
+  Separator,
+  List,
+  LargeArrowIconNext,
+  FacebookButtonIcon,
+  TwitterButtonIcon,
+  InstagramButtonIcon,
+  YoutubeButtonIcon,
+  EnvelopeIcon,
+  FloatingMenu,
+} from 'kitten'
+
+import Footer from 'kitten/karl/pages/footer'
+
+const StyledAvatar = styled.img`
+  --avatar-size: ${pxToRem(100)};
+
+  @media ${mq.tabletAndDesktop} {
+    --avatar-size: ${pxToRem(170)};
+  }
+
+  box-sizing: border-box;
+  object-fit: cover;
+  object-position: center;
+  width: var(--avatar-size);
+  height: var(--avatar-size);
+  border-radius: var(--avatar-size);
+  border: ${pxToRem(10)} solid var(--color-grey-000);
+`
+
+const StyledCounter = styled.div`
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+
+  .kiss-ProjectPage-hero-number {
+    ${TYPOGRAPHY.fontStyles.bold}
+    font-size: ${pxToRem(28)};
+    letter-spacing: -0.015em;
+    line-height: 1;
+  }
+  .kiss-ProjectPage-hero-text {
+    ${TYPOGRAPHY.fontStyles.light}
+    font-size: ${pxToRem(12)};
+    line-height: 1;
+  }
+
+  @media ${mq.tabletAndDesktop} {
+    .kiss-ProjectPage-hero-number {
+      font-size: ${pxToRem(46)};
+      letter-spacing: -0.02em;
+    }
+    .kiss-ProjectPage-hero-text {
+      font-size: ${pxToRem(20)};
+      letter-spacing: -0.01em;
+    }
+  }
+`
+
+const StyledGridCol = styled(GridCol)`
+  display: flex;
+  gap: ${pxToRem(20)};
+  padding: 0 ${pxToRem(20)} !important;
+
+  svg {
+    margin-top: ${pxToRem(5)};
+    flex: 0 0 ${pxToRem(20)};
+    max-width: ${pxToRem(20)};
+  }
+`
+
+const GlobalStyle = createGlobalStyle`
+  @media ${mq.mobile} {
+    .k-HeaderNav--isStuck + * .k-HeroLayout__sticky.k-HeroLayout__menu {
+      top: ${pxToRem(HEADER_HEIGHT)} !important;
+    }
+  }
+
+  @media ${mq.tabletAndDesktop} {
+    .k-HeaderNav--isStuck + * .k-HeroLayout__sticky__inside {
+      top: ${pxToRem(HEADER_HEIGHT + 20)} !important;
+    }
+  }
+`
+
+export default {
+  title: 'Layout/HeroLayout',
+  component: HeroLayout,
+  parameters: {
+    docs: {
+      page: () => <DocsPage filepath={__filename} importString="HeroLayout" />,
+    },
+  },
+  args: {
+    fullWidth: false,
+  },
+  argTypes: {
+    fullWidth: { control: 'boolean' },
+  },
+}
+
+export const Default = args => (
+  <>
+    <GlobalStyle />
+    <HeaderNav
+      id="header_id"
+      quickAccessProps={{
+        href: '#mainContent',
+        text: 'Aller au contenu principal',
+        zIndex: 300,
+      }}
+      size="small"
+      borderStyle="border"
+      zIndexConfig={{
+        header: 102,
+        headerWithOpenMenu: 105,
+      }}
+    >
+      <HeaderNav.Logo href="#">
+        <KissKissBankBankLogo height="25" className="k-u-margin-left-double" />
+      </HeaderNav.Logo>
+    </HeaderNav>
+    <HeroLayout
+      {...args}
+      hero_bg={<img src={`/kitten-${Math.floor(Math.random() * 10)}.jpg`} />}
+      hero_header={
+        <a href="#" className="k-u-block">
+          <StyledAvatar
+            src={`/kitten-${9 - Math.floor(Math.random() * 10)}.jpg`}
+            alt="Nom du porteur de projet"
+          />
+        </a>
+      }
+      hero_content={
+        <FlexWrapper className="k-u-flex-alignItems-center" gap={30}>
+          <Title noMargin modifier="primary">
+            Magna Condimentum Mollis Tortor
+          </Title>
+          <Text
+            tag="p"
+            weight="regular"
+            size="giant"
+            className="k-u-margin-none k-u-letter-spacing-10"
+          >
+            Duis mollis, est non commodo luctus, nisi erat Textorttitor ligula,
+            eget lacinia odio sem nec elit.
+          </Text>
+          <StyledCounter>
+            <span className="kiss-ProjectPage-hero-number">538</span>
+            <span className="kiss-ProjectPage-hero-text">abonnés</span>
+          </StyledCounter>
+          <FlexWrapper direction="row" gap={5}>
+            <Button fit="content" size="tiny" rounded>
+              Partager
+            </Button>
+            <Button fit="content" size="tiny" rounded>
+              Ajouter aux favoris
+            </Button>
+            <Button fit="content" size="tiny" modifier="helium" rounded>
+              S’abonner
+            </Button>
+          </FlexWrapper>
+        </FlexWrapper>
+      }
+      menu={
+        <FloatingMenu style={{ margin: '0 -10px' }}>
+          <FloatingMenu.Item href="#">Abonnements</FloatingMenu.Item>
+          <FloatingMenu.Item href="#" isActive>
+            À propos
+          </FloatingMenu.Item>
+          <FloatingMenu.Item href="#">Actualités</FloatingMenu.Item>
+          <FloatingMenu.Item href="#">FAQ</FloatingMenu.Item>
+        </FloatingMenu>
+      }
+      aside={
+        <FlexWrapper gap={20}>
+          <div>
+            <Text
+              tag="h3"
+              size="micro"
+              transform="uppercase"
+              className="k-u-margin-top-none k-u-margin-bottom-single k-u-letter-spacing-10"
+            >
+              Nous suivre
+            </Text>
+            <FlexWrapper direction="row" gap={10}>
+              <a href="#">
+                <FacebookButtonIcon rounded size="tiny" />
+              </a>
+              <a href="#">
+                <TwitterButtonIcon rounded size="tiny" />
+              </a>
+              <a href="#">
+                <InstagramButtonIcon rounded size="tiny" />
+              </a>
+              <a href="#">
+                <YoutubeButtonIcon rounded size="tiny" />
+              </a>
+            </FlexWrapper>
+          </div>
+
+          <div>
+            <Text
+              tag="h3"
+              size="micro"
+              transform="uppercase"
+              className="k-u-margin-top-none k-u-margin-bottom-noneHalf k-u-letter-spacing-10"
+            >
+              Site internet
+            </Text>
+            <div>
+              <Button fit="content" rounded size="micro">
+                www.mad-movies.com
+              </Button>
+            </div>
+          </div>
+
+          <div>
+            <Text
+              tag="h3"
+              size="micro"
+              transform="uppercase"
+              className="k-u-margin-top-none k-u-margin-bottom-noneHalf k-u-letter-spacing-10"
+            >
+              Contact
+            </Text>
+            <div>
+              <Button fit="content" rounded size="tiny">
+                <EnvelopeIcon />
+                <span>Nous écrire</span>
+              </Button>
+            </div>
+          </div>
+        </FlexWrapper>
+      }
+      promo={
+        <>
+          <Grid>
+            <GridCol
+              col="12"
+              col-s="6"
+              col-l="4"
+              offset-l="2"
+              className="k-u-align-center"
+            >
+              <Title tag="h2" noMargin modifier="tertiary">
+                Abonnez-vous
+              </Title>
+              <Text>
+                à partir de{' '}
+                <strong className="k-u-weight-bold">12&nbsp;€</strong> par mois
+              </Text>
+            </GridCol>
+            <GridCol col="12" col-s="6" col-l="4">
+              <Text
+                tag="h3"
+                weight="bold"
+                transform="uppercase"
+                size="tiny"
+                cssColor="var(--color-primary-500)"
+                className="k-u-letter-spacing-10 k-u-margin-top-none k-u-margin-bottom-singleHalf"
+              >
+                Pourquoi s’abonner&nbsp;?
+              </Text>
+              <List
+                markerElement={
+                  <LargeArrowIconNext
+                    color="var(--color-primary-500)"
+                    width="10"
+                  />
+                }
+              >
+                <Text size="big">Contribuez à faire vivre ce projet</Text>
+                <Text size="big">Recevez un contenu exclusif</Text>
+                <Text size="big">
+                  Faites partie du cercle privilégié d’abonnés
+                </Text>
+              </List>
+            </GridCol>
+          </Grid>
+          <Separator darker className="k-u-margin-vertical-double" />
+          <Grid>
+            <GridCol col="12">
+              <Text
+                tag="h3"
+                weight="bold"
+                transform="uppercase"
+                size="tiny"
+                cssColor="var(--color-primary-500)"
+                className="k-u-letter-spacing-10 k-u-margin-top-none k-u-margin-bottom-singleHalf"
+              >
+                Comment ça marche&nbsp;?
+              </Text>
+            </GridCol>
+            <StyledGridCol col="12" col-s="6" col-l="3">
+              <HeartIconNext width="30" />
+              <Text>
+                Sélectionnez un abonnement ou faites un don récurrent du montant
+                de votre choix
+              </Text>
+            </StyledGridCol>
+            <StyledGridCol col="12" col-s="6" col-l="3">
+              <HeartIconNext width="30" />
+              <Text>
+                Validez votre abonnement via notre formulaire de paiement
+                sécurisé
+              </Text>
+            </StyledGridCol>
+            <StyledGridCol col="12" col-s="6" col-l="3">
+              <HeartIconNext width="30" />
+              <Text>
+                Recevez chaque mois ou chaque trimestre le contenu de votre
+                abonnement
+              </Text>
+            </StyledGridCol>
+            <StyledGridCol col="12" col-s="6" col-l="3">
+              <HeartIconNext width="30" />
+              <Text>
+                Résiliez votre abonnement à tout moment depuis votre compte
+              </Text>
+            </StyledGridCol>
+          </Grid>
+        </>
+      }
+    >
+      <p className="k-u-margin-top-none">
+        Sed posuere consectetur est at lobortis. Cras mattis consectetur purus
+        sit amet fermentum. Curabitur blandit tempus porttitor. Nullam quis
+        risus eget urna mollis ornare vel eu leo.
+      </p>
+      <p>
+        Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+        Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Morbi leo
+        risus, porta ac consectetur ac, vestibulum at eros. Fusce dapibus,
+        tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum
+        massa justo sit amet risus. Etiam porta sem malesuada magna mollis
+        euismod. Curabitur blandit tempus porttitor.
+      </p>
+      <p>
+        Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+        Nullam quis risus eget urna mollis ornare vel eu leo. Praesent commodo
+        cursus magna, vel scelerisque nisl consectetur et. Sed posuere
+        consectetur est at lobortis. Vestibulum id ligula porta felis euismod
+        semper.
+      </p>
+      <p>
+        Sed posuere consectetur est at lobortis. Cras mattis consectetur purus
+        sit amet fermentum. Curabitur blandit tempus porttitor. Nullam quis
+        risus eget urna mollis ornare vel eu leo.
+      </p>
+      <p>
+        Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+        Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Morbi leo
+        risus, porta ac consectetur ac, vestibulum at eros. Fusce dapibus,
+        tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum
+        massa justo sit amet risus. Etiam porta sem malesuada magna mollis
+        euismod. Curabitur blandit tempus porttitor.
+      </p>
+      <p>
+        Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+        Nullam quis risus eget urna mollis ornare vel eu leo. Praesent commodo
+        cursus magna, vel scelerisque nisl consectetur et. Sed posuere
+        consectetur est at lobortis. Vestibulum id ligula porta felis euismod
+        semper.
+      </p>
+      <p>
+        Sed posuere consectetur est at lobortis. Cras mattis consectetur purus
+        sit amet fermentum. Curabitur blandit tempus porttitor. Nullam quis
+        risus eget urna mollis ornare vel eu leo.
+      </p>
+      <p>
+        Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+        Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Morbi leo
+        risus, porta ac consectetur ac, vestibulum at eros. Fusce dapibus,
+        tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum
+        massa justo sit amet risus. Etiam porta sem malesuada magna mollis
+        euismod. Curabitur blandit tempus porttitor.
+      </p>
+      <p>
+        Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+        Nullam quis risus eget urna mollis ornare vel eu leo. Praesent commodo
+        cursus magna, vel scelerisque nisl consectetur et. Sed posuere
+        consectetur est at lobortis. Vestibulum id ligula porta felis euismod
+        semper.
+      </p>
+      <p>
+        Sed posuere consectetur est at lobortis. Cras mattis consectetur purus
+        sit amet fermentum. Curabitur blandit tempus porttitor. Nullam quis
+        risus eget urna mollis ornare vel eu leo.
+      </p>
+      <p>
+        Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+        Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Morbi leo
+        risus, porta ac consectetur ac, vestibulum at eros. Fusce dapibus,
+        tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum
+        massa justo sit amet risus. Etiam porta sem malesuada magna mollis
+        euismod. Curabitur blandit tempus porttitor.
+      </p>
+      <p className="k-u-margin-bottom-none">
+        Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+        Nullam quis risus eget urna mollis ornare vel eu leo. Praesent commodo
+        cursus magna, vel scelerisque nisl consectetur et. Sed posuere
+        consectetur est at lobortis. Vestibulum id ligula porta felis euismod
+        semper.
+      </p>
+    </HeroLayout>
+    <Footer className="k-u-margin-top-quintuple" />
+  </>
+)